requirements.txt
langchain_core
python-dotenv
streamlit
langchain_community
chromadb
faiss-cpu
google.generativeai
langchain_google_genai
beautifulsoup4
bs4